refactor: Change the data type to the fields productId and amount

+++ src/main/java/io/shoppingcard/business/discount/CalculateDiscount.java
@@ -2,5 +2,5 @@ package io.shoppingcard.business.discount;
 
 @FunctionalInterface
 public interface CalculateDiscount {
-    Float apply(Integer productId, Float value);
+    Float apply(Long productId, Long value);
 }
