+++ src/main/java/io/shoppingcard/business/discount/GetProductDiscount.java
@@ -2,5 +2,5 @@ package io.shoppingcard.business.discount;
 
 @FunctionalInterface
 public interface GetProductDiscount {
-    Float getPercent(Integer productId);
+    Float getPercent(Long productId);
 }
